+++ src/main/scala/com/github/gchudnov/mtg/internal/BasicOps.scala
@@ -6,33 +6,35 @@ import com.github.gchudnov.mtg.Interval
 /**
  * Basic Interval Operations:
  *   - Intersection
+ *   - Span
+ *
+ * {{{
+ *             b-           b+
+ * |           |------------|
+ * |           .            .
+ * | |------|  .            .  |------|
+ * |           .            .
+ * |    |------|            |------|
+ * |           .            .
+ * |       |---+--|      |--+---|
+ * |           .            .
+ * |           |------|     .
+ * |           .            .
+ * |           .  |------|  .
+ * |           .            .
+ * |           .     |------|
+ * |           .            .
+ * |           .            .
+ * }}}
  */
 private[mtg] transparent trait BasicOps[+T]:
   a: Interval[T] =>
 
   /**
-   * Intersection of two Intervals: A ∩ B
+   * Intersection of two intervals
    *
-   * A & B
-   *
-   * {{{
-   *             b-           b+
-   * |           |------------|
-   * |           .            .
-   * | |------|  .            .  |------|
-   * |           .            .
-   * |    |------|            |------|
-   * |           .            .
-   * |       |---+--|      |--+---|
-   * |           .            .
-   * |           |------|     .
-   * |           .            .
-   * |           .  |------|  .
-   * |           .            .
-   * |           .     |------|
-   * |           .            .
-   * |           .            .
-   * }}}
+   *   - A ∩ B
+   *   - A & B
    *
    * {{{
    *          | ∅        if (A <b|B> B)       | b, B
@@ -41,9 +43,22 @@ private[mtg] transparent trait BasicOps[+T]:
    *          | [b-, a+] if (A <m|o|F> B)     | m, o, F
    *          | [b-, b+] if (A <D> B)         | D
    *
-   * A ∩ B := | max(a−, b−), min(a+, b+) |
+   * A ∩ B := | max(a-, b-), min(a+, b+) |
    * }}}
    */
   final def intersection[T1 >: T](b: Interval[T1])(using ordT: Ordering[Boundary[T1]]): Interval[T1] =
     if a.isEmpty || b.isEmpty then Interval.empty[T]
     else Interval.make(ordT.max(a.left, b.left), ordT.min(a.right, b.right))
+
+  /**
+   * Span of two intervals
+   *
+   *   - A # B
+   *
+   * {{{
+   * A # B := | min(a-, b-), max(a+, b+) |
+   * }}}
+   */
+  final def span[T1 >: T](b: Interval[T1])(using ordT: Ordering[Boundary[T1]]): Interval[T1] =
+    if a.isEmpty || b.isEmpty then Interval.empty[T]
+    else Interval.make(ordT.min(a.left, b.left), ordT.max(a.right, b.right))
